Started adding project configurations for release as well

+++ TrustHubWin/UserspaceTest/UserspaceTest.cpp
@@ -78,7 +78,7 @@ int main(int argc, char* argv[]) {
 	printf("Process Path Size: %lx\r\n", processPathSize);
 	cursor += sizeof(UINT32);
 
-	for (int i = 0; i < processPathSize; i++) {
+	for (int i = 0; i < (int)processPathSize; i++) {
 		printf("%s%02x ", (!(i % 16) ? "\r\n" : (!(i % 8) ? "   " : (!(i % 4) ? " " : ""))), *cursor);
 		cursor++;
 	}
@@ -88,7 +88,7 @@ int main(int argc, char* argv[]) {
 	printf("Client Hello Size: %lx\r\n", clientHelloSize);
 	cursor += sizeof(UINT32);
 
-	for (int i = 0; i < clientHelloSize; i++) {
+	for (int i = 0; i < (int)clientHelloSize; i++) {
 		printf("%s%02x ", (!(i % 16) ? "\r\n" : (!(i % 8) ? "   " : (!(i % 4) ? " " : ""))), *cursor);
 		cursor++;
 	}
@@ -98,7 +98,7 @@ int main(int argc, char* argv[]) {
 	printf("Server Hello Size: %lx\r\n", serverHelloSize);
 	cursor += sizeof(UINT32);
 
-	for (int i = 0; i < serverHelloSize; i++) {
+	for (int i = 0; i < (int)serverHelloSize; i++) {
 		printf("%s%02x ", (!(i % 16) ? "\r\n" : (!(i % 8) ? "   " : (!(i % 4) ? " " : ""))), *cursor);
 		cursor++;
 	}
@@ -108,7 +108,7 @@ int main(int argc, char* argv[]) {
 	printf("Certificate Size: %lx\r\n", dataSize);
 	cursor += sizeof(UINT32);
 
-	for (int i = 0; i < dataSize; i++) {
+	for (int i = 0; i < (int)dataSize; i++) {
 		printf("%s%02x ", (!(i % 16) ? "\r\n" : (!(i % 8) ? "   " : (!(i % 4) ? " " : ""))), *cursor);
 		cursor++;
 	}
